Functions that have been renamed, superseded, or are no longer working

Generally these functions have been superseded or renamed. Upon using them, a message is provided directing the user to the new function. However, at this stage of package development, we generally clear older defunct functions at each minor release, and so you are strongly encouraged to use the new functions/names/syntax wherever possible and update your scripts accordingly.

Usage

is_element(.data)

md_intersect(.data)

md_union(.data)

md_multiset(.data)

Functions

is_element(): Deprecated on 2023-08-25.

md_intersect(): Deprecated on 2023-08-25.

md_union(): Deprecated on 2023-08-25.

md_multiset(): Deprecated on 2023-08-25.