Expand messy dates to lists of dates

These functions expand on unspecified dates, ranges and on sets of dates. Uncertain dates may include several possible dates. The function "opens" these values to include all the possible dates contained in uncertain dates.

expand(x)

# S3 method for messydt
expand(x)

Arguments

x
A messydt object

Methods (by class)

messydt: Expanding messydates

Examples

d <- as_messydate(c("2001-01-01", "2001-01", "2001", "2001-01-01..2001-02-02",
        "{2001-01-01,2001-02-02}", "{2001-01,2001-02-02}"))
expand(d)

#> [[1]]
#> [1] "2001-01-01"
#> 
#> [[2]]
#>  [1] "2001-01-01" "2001-01-02" "2001-01-03" "2001-01-04" "2001-01-05"
#>  [6] "2001-01-06" "2001-01-07" "2001-01-08" "2001-01-09" "2001-01-10"
#> [11] "2001-01-11" "2001-01-12" "2001-01-13" "2001-01-14" "2001-01-15"
#> [16] "2001-01-16" "2001-01-17" "2001-01-18" "2001-01-19" "2001-01-20"
#> [21] "2001-01-21" "2001-01-22" "2001-01-23" "2001-01-24" "2001-01-25"
#> [26] "2001-01-26" "2001-01-27" "2001-01-28" "2001-01-29" "2001-01-30"
#> [31] "2001-01-31"
#> 
#> [[3]]
#>   [1] "2001-01-01" "2001-01-02" "2001-01-03" "2001-01-04" "2001-01-05"
#>   [6] "2001-01-06" "2001-01-07" "2001-01-08" "2001-01-09" "2001-01-10"
#>  [11] "2001-01-11" "2001-01-12" "2001-01-13" "2001-01-14" "2001-01-15"
#>  [16] "2001-01-16" "2001-01-17" "2001-01-18" "2001-01-19" "2001-01-20"
#>  [21] "2001-01-21" "2001-01-22" "2001-01-23" "2001-01-24" "2001-01-25"
#>  [26] "2001-01-26" "2001-01-27" "2001-01-28" "2001-01-29" "2001-01-30"
#>  [31] "2001-01-31" "2001-02-01" "2001-02-02" "2001-02-03" "2001-02-04"
#>  [36] "2001-02-05" "2001-02-06" "2001-02-07" "2001-02-08" "2001-02-09"
#>  [41] "2001-02-10" "2001-02-11" "2001-02-12" "2001-02-13" "2001-02-14"
#>  [46] "2001-02-15" "2001-02-16" "2001-02-17" "2001-02-18" "2001-02-19"
#>  [51] "2001-02-20" "2001-02-21" "2001-02-22" "2001-02-23" "2001-02-24"
#>  [56] "2001-02-25" "2001-02-26" "2001-02-27" "2001-02-28" "2001-03-01"
#>  [61] "2001-03-02" "2001-03-03" "2001-03-04" "2001-03-05" "2001-03-06"
#>  [66] "2001-03-07" "2001-03-08" "2001-03-09" "2001-03-10" "2001-03-11"
#>  [71] "2001-03-12" "2001-03-13" "2001-03-14" "2001-03-15" "2001-03-16"
#>  [76] "2001-03-17" "2001-03-18" "2001-03-19" "2001-03-20" "2001-03-21"
#>  [81] "2001-03-22" "2001-03-23" "2001-03-24" "2001-03-25" "2001-03-26"
#>  [86] "2001-03-27" "2001-03-28" "2001-03-29" "2001-03-30" "2001-03-31"
#>  [91] "2001-04-01" "2001-04-02" "2001-04-03" "2001-04-04" "2001-04-05"
#>  [96] "2001-04-06" "2001-04-07" "2001-04-08" "2001-04-09" "2001-04-10"
#> [101] "2001-04-11" "2001-04-12" "2001-04-13" "2001-04-14" "2001-04-15"
#> [106] "2001-04-16" "2001-04-17" "2001-04-18" "2001-04-19" "2001-04-20"
#> [111] "2001-04-21" "2001-04-22" "2001-04-23" "2001-04-24" "2001-04-25"
#> [116] "2001-04-26" "2001-04-27" "2001-04-28" "2001-04-29" "2001-04-30"
#> [121] "2001-05-01" "2001-05-02" "2001-05-03" "2001-05-04" "2001-05-05"
#> [126] "2001-05-06" "2001-05-07" "2001-05-08" "2001-05-09" "2001-05-10"
#> [131] "2001-05-11" "2001-05-12" "2001-05-13" "2001-05-14" "2001-05-15"
#> [136] "2001-05-16" "2001-05-17" "2001-05-18" "2001-05-19" "2001-05-20"
#> [141] "2001-05-21" "2001-05-22" "2001-05-23" "2001-05-24" "2001-05-25"
#> [146] "2001-05-26" "2001-05-27" "2001-05-28" "2001-05-29" "2001-05-30"
#> [151] "2001-05-31" "2001-06-01" "2001-06-02" "2001-06-03" "2001-06-04"
#> [156] "2001-06-05" "2001-06-06" "2001-06-07" "2001-06-08" "2001-06-09"
#> [161] "2001-06-10" "2001-06-11" "2001-06-12" "2001-06-13" "2001-06-14"
#> [166] "2001-06-15" "2001-06-16" "2001-06-17" "2001-06-18" "2001-06-19"
#> [171] "2001-06-20" "2001-06-21" "2001-06-22" "2001-06-23" "2001-06-24"
#> [176] "2001-06-25" "2001-06-26" "2001-06-27" "2001-06-28" "2001-06-29"
#> [181] "2001-06-30" "2001-07-01" "2001-07-02" "2001-07-03" "2001-07-04"
#> [186] "2001-07-05" "2001-07-06" "2001-07-07" "2001-07-08" "2001-07-09"
#> [191] "2001-07-10" "2001-07-11" "2001-07-12" "2001-07-13" "2001-07-14"
#> [196] "2001-07-15" "2001-07-16" "2001-07-17" "2001-07-18" "2001-07-19"
#> [201] "2001-07-20" "2001-07-21" "2001-07-22" "2001-07-23" "2001-07-24"
#> [206] "2001-07-25" "2001-07-26" "2001-07-27" "2001-07-28" "2001-07-29"
#> [211] "2001-07-30" "2001-07-31" "2001-08-01" "2001-08-02" "2001-08-03"
#> [216] "2001-08-04" "2001-08-05" "2001-08-06" "2001-08-07" "2001-08-08"
#> [221] "2001-08-09" "2001-08-10" "2001-08-11" "2001-08-12" "2001-08-13"
#> [226] "2001-08-14" "2001-08-15" "2001-08-16" "2001-08-17" "2001-08-18"
#> [231] "2001-08-19" "2001-08-20" "2001-08-21" "2001-08-22" "2001-08-23"
#> [236] "2001-08-24" "2001-08-25" "2001-08-26" "2001-08-27" "2001-08-28"
#> [241] "2001-08-29" "2001-08-30" "2001-08-31" "2001-09-01" "2001-09-02"
#> [246] "2001-09-03" "2001-09-04" "2001-09-05" "2001-09-06" "2001-09-07"
#> [251] "2001-09-08" "2001-09-09" "2001-09-10" "2001-09-11" "2001-09-12"
#> [256] "2001-09-13" "2001-09-14" "2001-09-15" "2001-09-16" "2001-09-17"
#> [261] "2001-09-18" "2001-09-19" "2001-09-20" "2001-09-21" "2001-09-22"
#> [266] "2001-09-23" "2001-09-24" "2001-09-25" "2001-09-26" "2001-09-27"
#> [271] "2001-09-28" "2001-09-29" "2001-09-30" "2001-10-01" "2001-10-02"
#> [276] "2001-10-03" "2001-10-04" "2001-10-05" "2001-10-06" "2001-10-07"
#> [281] "2001-10-08" "2001-10-09" "2001-10-10" "2001-10-11" "2001-10-12"
#> [286] "2001-10-13" "2001-10-14" "2001-10-15" "2001-10-16" "2001-10-17"
#> [291] "2001-10-18" "2001-10-19" "2001-10-20" "2001-10-21" "2001-10-22"
#> [296] "2001-10-23" "2001-10-24" "2001-10-25" "2001-10-26" "2001-10-27"
#> [301] "2001-10-28" "2001-10-29" "2001-10-30" "2001-10-31" "2001-11-01"
#> [306] "2001-11-02" "2001-11-03" "2001-11-04" "2001-11-05" "2001-11-06"
#> [311] "2001-11-07" "2001-11-08" "2001-11-09" "2001-11-10" "2001-11-11"
#> [316] "2001-11-12" "2001-11-13" "2001-11-14" "2001-11-15" "2001-11-16"
#> [321] "2001-11-17" "2001-11-18" "2001-11-19" "2001-11-20" "2001-11-21"
#> [326] "2001-11-22" "2001-11-23" "2001-11-24" "2001-11-25" "2001-11-26"
#> [331] "2001-11-27" "2001-11-28" "2001-11-29" "2001-11-30" "2001-12-01"
#> [336] "2001-12-02" "2001-12-03" "2001-12-04" "2001-12-05" "2001-12-06"
#> [341] "2001-12-07" "2001-12-08" "2001-12-09" "2001-12-10" "2001-12-11"
#> [346] "2001-12-12" "2001-12-13" "2001-12-14" "2001-12-15" "2001-12-16"
#> [351] "2001-12-17" "2001-12-18" "2001-12-19" "2001-12-20" "2001-12-21"
#> [356] "2001-12-22" "2001-12-23" "2001-12-24" "2001-12-25" "2001-12-26"
#> [361] "2001-12-27" "2001-12-28" "2001-12-29" "2001-12-30" "2001-12-31"
#> 
#> [[4]]
#>  [1] "2001-01-01" "2001-01-02" "2001-01-03" "2001-01-04" "2001-01-05"
#>  [6] "2001-01-06" "2001-01-07" "2001-01-08" "2001-01-09" "2001-01-10"
#> [11] "2001-01-11" "2001-01-12" "2001-01-13" "2001-01-14" "2001-01-15"
#> [16] "2001-01-16" "2001-01-17" "2001-01-18" "2001-01-19" "2001-01-20"
#> [21] "2001-01-21" "2001-01-22" "2001-01-23" "2001-01-24" "2001-01-25"
#> [26] "2001-01-26" "2001-01-27" "2001-01-28" "2001-01-29" "2001-01-30"
#> [31] "2001-01-31" "2001-02-01" "2001-02-02"
#> 
#> [[5]]
#> [1] "2001-01-01" "2001-02-02"
#> 
#> [[6]]
#>  [1] "2001-01-01" "2001-01-02" "2001-01-03" "2001-01-04" "2001-01-05"
#>  [6] "2001-01-06" "2001-01-07" "2001-01-08" "2001-01-09" "2001-01-10"
#> [11] "2001-01-11" "2001-01-12" "2001-01-13" "2001-01-14" "2001-01-15"
#> [16] "2001-01-16" "2001-01-17" "2001-01-18" "2001-01-19" "2001-01-20"
#> [21] "2001-01-21" "2001-01-22" "2001-01-23" "2001-01-24" "2001-01-25"
#> [26] "2001-01-26" "2001-01-27" "2001-01-28" "2001-01-29" "2001-01-30"
#> [31] "2001-01-31" "2001-02-02"
#>