Expand messy dates to lists of dates

These functions expand on date ranges, sets of dates, and unspecified or approximate dates (annotated with '..', '', 'XX' or '~'). As these messydates may refer to several possible dates, the function "opens" these values to reveal a vector of all the possible dates implied. Imprecise dates (dates only containing information on year and/or month) are also expanded to include possible dates within that year and/or month. The function removes the annotation from dates with unreliable sources ('?'), before being expanded normally as though they were incomplete.

expand(x, approx_range = 0)

Arguments

x

A mdate object. If not an 'mdate' object, conversion is handled first with ´as_messydate()´.

approx_range

Range to expand approximate dates, or date components, annotated with '~', by default 0. That is, removes signs for approximate dates and treats these dates as precise dates. If 3, for example, adds 3 days for day approximation, 3 months for month approximation, 3 years for year/whole date approximation, 3 years and 3 months for year-month approximation, and 3 months and 3 days for month-day approximation.

Value

A list of dates, including all dates in each range or set.

Examples

d <- as_messydate(c("2008-03-25", "-2012-02-27", "2001-01?", "~2001",
"2001-01-01..2001-02-02", "{2001-01-01,2001-02-02}", "{2001-01,2001-02-02}",
"2008-XX-31", "..2002-02-03", "2001-01-03..", "28 BC"))
expand(d)
#> [[1]]
#> [1] "2008-03-25"
#> 
#> [[2]]
#> [1] "-2012-02-27"
#> 
#> [[3]]
#>  [1] "2001-01-01" "2001-01-02" "2001-01-03" "2001-01-04" "2001-01-05"
#>  [6] "2001-01-06" "2001-01-07" "2001-01-08" "2001-01-09" "2001-01-10"
#> [11] "2001-01-11" "2001-01-12" "2001-01-13" "2001-01-14" "2001-01-15"
#> [16] "2001-01-16" "2001-01-17" "2001-01-18" "2001-01-19" "2001-01-20"
#> [21] "2001-01-21" "2001-01-22" "2001-01-23" "2001-01-24" "2001-01-25"
#> [26] "2001-01-26" "2001-01-27" "2001-01-28" "2001-01-29" "2001-01-30"
#> [31] "2001-01-31"
#> 
#> [[4]]
#>   [1] "2001-01-01" "2001-01-02" "2001-01-03" "2001-01-04" "2001-01-05"
#>   [6] "2001-01-06" "2001-01-07" "2001-01-08" "2001-01-09" "2001-01-10"
#>  [11] "2001-01-11" "2001-01-12" "2001-01-13" "2001-01-14" "2001-01-15"
#>  [16] "2001-01-16" "2001-01-17" "2001-01-18" "2001-01-19" "2001-01-20"
#>  [21] "2001-01-21" "2001-01-22" "2001-01-23" "2001-01-24" "2001-01-25"
#>  [26] "2001-01-26" "2001-01-27" "2001-01-28" "2001-01-29" "2001-01-30"
#>  [31] "2001-01-31" "2001-02-01" "2001-02-02" "2001-02-03" "2001-02-04"
#>  [36] "2001-02-05" "2001-02-06" "2001-02-07" "2001-02-08" "2001-02-09"
#>  [41] "2001-02-10" "2001-02-11" "2001-02-12" "2001-02-13" "2001-02-14"
#>  [46] "2001-02-15" "2001-02-16" "2001-02-17" "2001-02-18" "2001-02-19"
#>  [51] "2001-02-20" "2001-02-21" "2001-02-22" "2001-02-23" "2001-02-24"
#>  [56] "2001-02-25" "2001-02-26" "2001-02-27" "2001-02-28" "2001-03-01"
#>  [61] "2001-03-02" "2001-03-03" "2001-03-04" "2001-03-05" "2001-03-06"
#>  [66] "2001-03-07" "2001-03-08" "2001-03-09" "2001-03-10" "2001-03-11"
#>  [71] "2001-03-12" "2001-03-13" "2001-03-14" "2001-03-15" "2001-03-16"
#>  [76] "2001-03-17" "2001-03-18" "2001-03-19" "2001-03-20" "2001-03-21"
#>  [81] "2001-03-22" "2001-03-23" "2001-03-24" "2001-03-25" "2001-03-26"
#>  [86] "2001-03-27" "2001-03-28" "2001-03-29" "2001-03-30" "2001-03-31"
#>  [91] "2001-04-01" "2001-04-02" "2001-04-03" "2001-04-04" "2001-04-05"
#>  [96] "2001-04-06" "2001-04-07" "2001-04-08" "2001-04-09" "2001-04-10"
#> [101] "2001-04-11" "2001-04-12" "2001-04-13" "2001-04-14" "2001-04-15"
#> [106] "2001-04-16" "2001-04-17" "2001-04-18" "2001-04-19" "2001-04-20"
#> [111] "2001-04-21" "2001-04-22" "2001-04-23" "2001-04-24" "2001-04-25"
#> [116] "2001-04-26" "2001-04-27" "2001-04-28" "2001-04-29" "2001-04-30"
#> [121] "2001-05-01" "2001-05-02" "2001-05-03" "2001-05-04" "2001-05-05"
#> [126] "2001-05-06" "2001-05-07" "2001-05-08" "2001-05-09" "2001-05-10"
#> [131] "2001-05-11" "2001-05-12" "2001-05-13" "2001-05-14" "2001-05-15"
#> [136] "2001-05-16" "2001-05-17" "2001-05-18" "2001-05-19" "2001-05-20"
#> [141] "2001-05-21" "2001-05-22" "2001-05-23" "2001-05-24" "2001-05-25"
#> [146] "2001-05-26" "2001-05-27" "2001-05-28" "2001-05-29" "2001-05-30"
#> [151] "2001-05-31" "2001-06-01" "2001-06-02" "2001-06-03" "2001-06-04"
#> [156] "2001-06-05" "2001-06-06" "2001-06-07" "2001-06-08" "2001-06-09"
#> [161] "2001-06-10" "2001-06-11" "2001-06-12" "2001-06-13" "2001-06-14"
#> [166] "2001-06-15" "2001-06-16" "2001-06-17" "2001-06-18" "2001-06-19"
#> [171] "2001-06-20" "2001-06-21" "2001-06-22" "2001-06-23" "2001-06-24"
#> [176] "2001-06-25" "2001-06-26" "2001-06-27" "2001-06-28" "2001-06-29"
#> [181] "2001-06-30" "2001-07-01" "2001-07-02" "2001-07-03" "2001-07-04"
#> [186] "2001-07-05" "2001-07-06" "2001-07-07" "2001-07-08" "2001-07-09"
#> [191] "2001-07-10" "2001-07-11" "2001-07-12" "2001-07-13" "2001-07-14"
#> [196] "2001-07-15" "2001-07-16" "2001-07-17" "2001-07-18" "2001-07-19"
#> [201] "2001-07-20" "2001-07-21" "2001-07-22" "2001-07-23" "2001-07-24"
#> [206] "2001-07-25" "2001-07-26" "2001-07-27" "2001-07-28" "2001-07-29"
#> [211] "2001-07-30" "2001-07-31" "2001-08-01" "2001-08-02" "2001-08-03"
#> [216] "2001-08-04" "2001-08-05" "2001-08-06" "2001-08-07" "2001-08-08"
#> [221] "2001-08-09" "2001-08-10" "2001-08-11" "2001-08-12" "2001-08-13"
#> [226] "2001-08-14" "2001-08-15" "2001-08-16" "2001-08-17" "2001-08-18"
#> [231] "2001-08-19" "2001-08-20" "2001-08-21" "2001-08-22" "2001-08-23"
#> [236] "2001-08-24" "2001-08-25" "2001-08-26" "2001-08-27" "2001-08-28"
#> [241] "2001-08-29" "2001-08-30" "2001-08-31" "2001-09-01" "2001-09-02"
#> [246] "2001-09-03" "2001-09-04" "2001-09-05" "2001-09-06" "2001-09-07"
#> [251] "2001-09-08" "2001-09-09" "2001-09-10" "2001-09-11" "2001-09-12"
#> [256] "2001-09-13" "2001-09-14" "2001-09-15" "2001-09-16" "2001-09-17"
#> [261] "2001-09-18" "2001-09-19" "2001-09-20" "2001-09-21" "2001-09-22"
#> [266] "2001-09-23" "2001-09-24" "2001-09-25" "2001-09-26" "2001-09-27"
#> [271] "2001-09-28" "2001-09-29" "2001-09-30" "2001-10-01" "2001-10-02"
#> [276] "2001-10-03" "2001-10-04" "2001-10-05" "2001-10-06" "2001-10-07"
#> [281] "2001-10-08" "2001-10-09" "2001-10-10" "2001-10-11" "2001-10-12"
#> [286] "2001-10-13" "2001-10-14" "2001-10-15" "2001-10-16" "2001-10-17"
#> [291] "2001-10-18" "2001-10-19" "2001-10-20" "2001-10-21" "2001-10-22"
#> [296] "2001-10-23" "2001-10-24" "2001-10-25" "2001-10-26" "2001-10-27"
#> [301] "2001-10-28" "2001-10-29" "2001-10-30" "2001-10-31" "2001-11-01"
#> [306] "2001-11-02" "2001-11-03" "2001-11-04" "2001-11-05" "2001-11-06"
#> [311] "2001-11-07" "2001-11-08" "2001-11-09" "2001-11-10" "2001-11-11"
#> [316] "2001-11-12" "2001-11-13" "2001-11-14" "2001-11-15" "2001-11-16"
#> [321] "2001-11-17" "2001-11-18" "2001-11-19" "2001-11-20" "2001-11-21"
#> [326] "2001-11-22" "2001-11-23" "2001-11-24" "2001-11-25" "2001-11-26"
#> [331] "2001-11-27" "2001-11-28" "2001-11-29" "2001-11-30" "2001-12-01"
#> [336] "2001-12-02" "2001-12-03" "2001-12-04" "2001-12-05" "2001-12-06"
#> [341] "2001-12-07" "2001-12-08" "2001-12-09" "2001-12-10" "2001-12-11"
#> [346] "2001-12-12" "2001-12-13" "2001-12-14" "2001-12-15" "2001-12-16"
#> [351] "2001-12-17" "2001-12-18" "2001-12-19" "2001-12-20" "2001-12-21"
#> [356] "2001-12-22" "2001-12-23" "2001-12-24" "2001-12-25" "2001-12-26"
#> [361] "2001-12-27" "2001-12-28" "2001-12-29" "2001-12-30" "2001-12-31"
#> 
#> [[5]]
#>  [1] "2001-01-01" "2001-01-02" "2001-01-03" "2001-01-04" "2001-01-05"
#>  [6] "2001-01-06" "2001-01-07" "2001-01-08" "2001-01-09" "2001-01-10"
#> [11] "2001-01-11" "2001-01-12" "2001-01-13" "2001-01-14" "2001-01-15"
#> [16] "2001-01-16" "2001-01-17" "2001-01-18" "2001-01-19" "2001-01-20"
#> [21] "2001-01-21" "2001-01-22" "2001-01-23" "2001-01-24" "2001-01-25"
#> [26] "2001-01-26" "2001-01-27" "2001-01-28" "2001-01-29" "2001-01-30"
#> [31] "2001-01-31" "2001-02-01" "2001-02-02"
#> 
#> [[6]]
#> [1] "2001-01-01" "2001-02-02"
#> 
#> [[7]]
#>  [1] "2001-01-01" "2001-01-02" "2001-01-03" "2001-01-04" "2001-01-05"
#>  [6] "2001-01-06" "2001-01-07" "2001-01-08" "2001-01-09" "2001-01-10"
#> [11] "2001-01-11" "2001-01-12" "2001-01-13" "2001-01-14" "2001-01-15"
#> [16] "2001-01-16" "2001-01-17" "2001-01-18" "2001-01-19" "2001-01-20"
#> [21] "2001-01-21" "2001-01-22" "2001-01-23" "2001-01-24" "2001-01-25"
#> [26] "2001-01-26" "2001-01-27" "2001-01-28" "2001-01-29" "2001-01-30"
#> [31] "2001-01-31" "2001-02-02"
#> 
#> [[8]]
#>  [1] "2008-01-31" "2008-02-28" "2008-03-31" "2008-04-30" "2008-05-31"
#>  [6] "2008-06-30" "2008-07-31" "2008-08-31" "2008-09-30" "2008-10-31"
#> [11] "2008-11-30" "2008-12-31"
#> 
#> [[9]]
#> [1] "2002-02-03"
#> 
#> [[10]]
#> [1] "2001-01-03"
#> 
#> [[11]]
#>   [1] "-28-01-01" "-28-01-02" "-28-01-03" "-28-01-04" "-28-01-05" "-28-01-06"
#>   [7] "-28-01-07" "-28-01-08" "-28-01-09" "-28-01-10" "-28-01-11" "-28-01-12"
#>  [13] "-28-01-13" "-28-01-14" "-28-01-15" "-28-01-16" "-28-01-17" "-28-01-18"
#>  [19] "-28-01-19" "-28-01-20" "-28-01-21" "-28-01-22" "-28-01-23" "-28-01-24"
#>  [25] "-28-01-25" "-28-01-26" "-28-01-27" "-28-01-28" "-28-01-29" "-28-01-30"
#>  [31] "-28-01-31" "-28-02-01" "-28-02-02" "-28-02-03" "-28-02-04" "-28-02-05"
#>  [37] "-28-02-06" "-28-02-07" "-28-02-08" "-28-02-09" "-28-02-10" "-28-02-11"
#>  [43] "-28-02-12" "-28-02-13" "-28-02-14" "-28-02-15" "-28-02-16" "-28-02-17"
#>  [49] "-28-02-18" "-28-02-19" "-28-02-20" "-28-02-21" "-28-02-22" "-28-02-23"
#>  [55] "-28-02-24" "-28-02-25" "-28-02-26" "-28-02-27" "-28-02-28" "-28-02-29"
#>  [61] "-28-03-01" "-28-03-02" "-28-03-03" "-28-03-04" "-28-03-05" "-28-03-06"
#>  [67] "-28-03-07" "-28-03-08" "-28-03-09" "-28-03-10" "-28-03-11" "-28-03-12"
#>  [73] "-28-03-13" "-28-03-14" "-28-03-15" "-28-03-16" "-28-03-17" "-28-03-18"
#>  [79] "-28-03-19" "-28-03-20" "-28-03-21" "-28-03-22" "-28-03-23" "-28-03-24"
#>  [85] "-28-03-25" "-28-03-26" "-28-03-27" "-28-03-28" "-28-03-29" "-28-03-30"
#>  [91] "-28-03-31" "-28-04-01" "-28-04-02" "-28-04-03" "-28-04-04" "-28-04-05"
#>  [97] "-28-04-06" "-28-04-07" "-28-04-08" "-28-04-09" "-28-04-10" "-28-04-11"
#> [103] "-28-04-12" "-28-04-13" "-28-04-14" "-28-04-15" "-28-04-16" "-28-04-17"
#> [109] "-28-04-18" "-28-04-19" "-28-04-20" "-28-04-21" "-28-04-22" "-28-04-23"
#> [115] "-28-04-24" "-28-04-25" "-28-04-26" "-28-04-27" "-28-04-28" "-28-04-29"
#> [121] "-28-04-30" "-28-05-01" "-28-05-02" "-28-05-03" "-28-05-04" "-28-05-05"
#> [127] "-28-05-06" "-28-05-07" "-28-05-08" "-28-05-09" "-28-05-10" "-28-05-11"
#> [133] "-28-05-12" "-28-05-13" "-28-05-14" "-28-05-15" "-28-05-16" "-28-05-17"
#> [139] "-28-05-18" "-28-05-19" "-28-05-20" "-28-05-21" "-28-05-22" "-28-05-23"
#> [145] "-28-05-24" "-28-05-25" "-28-05-26" "-28-05-27" "-28-05-28" "-28-05-29"
#> [151] "-28-05-30" "-28-05-31" "-28-06-01" "-28-06-02" "-28-06-03" "-28-06-04"
#> [157] "-28-06-05" "-28-06-06" "-28-06-07" "-28-06-08" "-28-06-09" "-28-06-10"
#> [163] "-28-06-11" "-28-06-12" "-28-06-13" "-28-06-14" "-28-06-15" "-28-06-16"
#> [169] "-28-06-17" "-28-06-18" "-28-06-19" "-28-06-20" "-28-06-21" "-28-06-22"
#> [175] "-28-06-23" "-28-06-24" "-28-06-25" "-28-06-26" "-28-06-27" "-28-06-28"
#> [181] "-28-06-29" "-28-06-30" "-28-07-01" "-28-07-02" "-28-07-03" "-28-07-04"
#> [187] "-28-07-05" "-28-07-06" "-28-07-07" "-28-07-08" "-28-07-09" "-28-07-10"
#> [193] "-28-07-11" "-28-07-12" "-28-07-13" "-28-07-14" "-28-07-15" "-28-07-16"
#> [199] "-28-07-17" "-28-07-18" "-28-07-19" "-28-07-20" "-28-07-21" "-28-07-22"
#> [205] "-28-07-23" "-28-07-24" "-28-07-25" "-28-07-26" "-28-07-27" "-28-07-28"
#> [211] "-28-07-29" "-28-07-30" "-28-07-31" "-28-08-01" "-28-08-02" "-28-08-03"
#> [217] "-28-08-04" "-28-08-05" "-28-08-06" "-28-08-07" "-28-08-08" "-28-08-09"
#> [223] "-28-08-10" "-28-08-11" "-28-08-12" "-28-08-13" "-28-08-14" "-28-08-15"
#> [229] "-28-08-16" "-28-08-17" "-28-08-18" "-28-08-19" "-28-08-20" "-28-08-21"
#> [235] "-28-08-22" "-28-08-23" "-28-08-24" "-28-08-25" "-28-08-26" "-28-08-27"
#> [241] "-28-08-28" "-28-08-29" "-28-08-30" "-28-08-31" "-28-09-01" "-28-09-02"
#> [247] "-28-09-03" "-28-09-04" "-28-09-05" "-28-09-06" "-28-09-07" "-28-09-08"
#> [253] "-28-09-09" "-28-09-10" "-28-09-11" "-28-09-12" "-28-09-13" "-28-09-14"
#> [259] "-28-09-15" "-28-09-16" "-28-09-17" "-28-09-18" "-28-09-19" "-28-09-20"
#> [265] "-28-09-21" "-28-09-22" "-28-09-23" "-28-09-24" "-28-09-25" "-28-09-26"
#> [271] "-28-09-27" "-28-09-28" "-28-09-29" "-28-09-30" "-28-10-01" "-28-10-02"
#> [277] "-28-10-03" "-28-10-04" "-28-10-05" "-28-10-06" "-28-10-07" "-28-10-08"
#> [283] "-28-10-09" "-28-10-10" "-28-10-11" "-28-10-12" "-28-10-13" "-28-10-14"
#> [289] "-28-10-15" "-28-10-16" "-28-10-17" "-28-10-18" "-28-10-19" "-28-10-20"
#> [295] "-28-10-21" "-28-10-22" "-28-10-23" "-28-10-24" "-28-10-25" "-28-10-26"
#> [301] "-28-10-27" "-28-10-28" "-28-10-29" "-28-10-30" "-28-10-31" "-28-11-01"
#> [307] "-28-11-02" "-28-11-03" "-28-11-04" "-28-11-05" "-28-11-06" "-28-11-07"
#> [313] "-28-11-08" "-28-11-09" "-28-11-10" "-28-11-11" "-28-11-12" "-28-11-13"
#> [319] "-28-11-14" "-28-11-15" "-28-11-16" "-28-11-17" "-28-11-18" "-28-11-19"
#> [325] "-28-11-20" "-28-11-21" "-28-11-22" "-28-11-23" "-28-11-24" "-28-11-25"
#> [331] "-28-11-26" "-28-11-27" "-28-11-28" "-28-11-29" "-28-11-30" "-28-12-01"
#> [337] "-28-12-02" "-28-12-03" "-28-12-04" "-28-12-05" "-28-12-06" "-28-12-07"
#> [343] "-28-12-08" "-28-12-09" "-28-12-10" "-28-12-11" "-28-12-12" "-28-12-13"
#> [349] "-28-12-14" "-28-12-15" "-28-12-16" "-28-12-17" "-28-12-18" "-28-12-19"
#> [355] "-28-12-20" "-28-12-21" "-28-12-22" "-28-12-23" "-28-12-24" "-28-12-25"
#> [361] "-28-12-26" "-28-12-27" "-28-12-28" "-28-12-29" "-28-12-30" "-28-12-31"
#>