Coercion from messy dates

These functions coerce objects of messydt class to common date classes such as Date, POSIXct, and POSIXlt. Since messydt objects can hold multiple individual dates, however, an additional function must be passed as an argument so that these functions know how to coerce resolve multiple dates into a single date.

For example, one might wish to use the earliest possible date in any ranges of dates (min), the latest possible date (max), some notion of a central tendency (mean, median, or modal), or even a random selection from amongst the candidate dates.

These functions then, building on expand() and the resolve functions, are particularly useful in converting back out of the messydt class for use with existing methods and models, especially for checking the robustness of results.

# S3 method for messydt
as.Date(x, ..., FUN)

# S3 method for messydt
as.POSIXct(x, ..., FUN)

# S3 method for messydt
as.POSIXlt(x, ..., FUN)

Arguments

x
A messydt object

...
Arguments passed on to the S3 generics.

FUN
A function that can be used to resolve expanded messy dates into a single date. For example, min(), max(), mean(), median(), modal(), and random().

Value

A date object of Date, POSIXct, or POSIXlt class

Examples

as.Date(as_messydate("2012-01"), min)

#> [1] "2012-01-01"

as.Date(as_messydate("2012-01"), mean)

#> [1] "2012-01-16"

as.Date(as_messydate("2012-01"), max)

#> [1] "2012-01-31"

as.Date(as_messydate("2012-01"), median)

#> [1] "2012-01-16"

as.Date(as_messydate("2012-01"), modal)

#> [1] "2012-01-01"

as.Date(as_messydate("2012-01"), random)

#> [1] "2012-01-13"